[rename, refactor] cleanXSS -> escapeForXSS, constants 디렉토리 생성, API 라우트 handleApiError 적용
