+++ src/app/api/auth/signout/route.ts
@@ -1,7 +1,7 @@
 import { NextResponse } from 'next/server';
 import { EXTERNAL_PATHS } from '@/lib/api/apiPaths';
-import { AxiosError } from 'axios';
 import { externalClient } from '@/lib/api/clientFetchers';
+import { handleApiError } from '@/lib/api/handleApiError';
 
 
 /**
@@ -13,7 +13,6 @@ export async function POST() {
         const response = await externalClient.post(EXTERNAL_PATHS.SIGN_OUT);
         return new NextResponse(JSON.stringify(response.data), { status: 200 });
     } catch (error) {
-        const err = error as AxiosError;
-        return new NextResponse(JSON.stringify({ error: err?.response?.data }), { status: 500 });
+        return handleApiError(error);
     }
 }
