+++ src/app/api/auth/signup/route.ts
@@ -1,7 +1,7 @@
 import { NextRequest, NextResponse } from 'next/server';
 import { EXTERNAL_PATHS } from '@/lib/api/apiPaths';
-import { AxiosError } from 'axios';
 import { externalClient } from '@/lib/api/clientFetchers';
+import { handleApiError } from '@/lib/api/handleApiError';
 
 /**
  * 회원가입
@@ -22,7 +22,6 @@ export async function POST(request: NextRequest) {
         })
         return new NextResponse(JSON.stringify(response.data), { status: 201 });
     } catch (error) {
-        const err = error as AxiosError;
-        return new NextResponse(JSON.stringify({ error: err?.response?.data }), { status: 500 });
+        return handleApiError(error);
     }
 }
