+++ src/utils/shared/excapeForXSS.ts
@@ -1,5 +1,5 @@
 /** 텍스트에 특수문자 치환 */
-export const cleanXSS = (text: string) => {
+export const excapeForXSS = (text: string) => {
     return text.replaceAll("<", "&lt;").replaceAll(">", "&gt;")
         .replaceAll("\\(", "& #40;").replaceAll("\\)", "& #41;")
         .replaceAll("'", "& #39;")
